Allow use of config_path for system config path.
The config_path for a system configuration setting can be specified in
system.xml. If this field is specified it overrides the path field for
the config option. Detect if config_path is present for a field and
return it rather than generating a path from it's position in the xml.